refactor: provisioning

+++ provision/web/provision.sh
@@ -26,6 +26,16 @@ dnf install -y unzip curl nano git util-linux-user \
   php-redis \
   composer
 
+# use full for redis-commander
+dnf install nodejs -y
+npm install -g redis-commander
+
+# Add more paths to the PATH variable
+if ! grep -q "export PATH=\$PATH:/usr/local/bin" ~/.bashrc; then
+    # If not present, append the line to .bashrc
+    echo 'export PATH=$PATH:/usr/local/bin' >> ~/.bashrc
+fi
+
 # Start and enable services
 systemctl start nginx
 systemctl enable nginx -q
@@ -40,43 +50,51 @@ systemctl disable firewalld -q
 #sudo firewall-cmd --permanent --add-service=http
 #sudo firewall-cmd --permanent --add-service=https
 
-# Set MySQL root password
-mysqladmin -u root password "root"
+FLAG_FILE="/etc/vagrant_provisioned"
 
-# Enable apache shell for development purpose
-chsh -s /bin/bash apache
+# only if file /var/www/html/typo3/ does not exist
+if [ ! -f "$FLAG_FILE" ]; then
 
-# Install TYPO3 dependencies
-echo "Installing TYPO3 specific dependencies"
-yum install -y ImageMagick GraphicsMagick ghostscript
+  # Set MySQL root password
+  mysqladmin -u root password "root"
 
-# Just for enabling composer caching
-mkdir /usr/share/httpd/{.composer,.vim}
-chown apache:apache /usr/share/httpd/{.composer,.vim}
+  # Enable apache shell for development purpose
+  chsh -s /bin/bash apache
 
-# SELinux serve files off Apache, resursive
-mkdir /var/www/html/typo3
-chown apache:apache /var/www/html/typo3
+  # Install TYPO3 dependencies
+  echo "Installing TYPO3 specific dependencies"
+  yum install -y ImageMagick GraphicsMagick ghostscript
 
-# Configure SELinux
-sudo chcon -t httpd_sys_rw_content_t /var/www/html/typo3 -R # change context
-sudo setsebool -P httpd_can_network_connect 1 # allow network connections
+  # Just for enabling composer caching
+  mkdir /usr/share/httpd/{.composer,.vim}
+  chown apache:apache /usr/share/httpd/{.composer,.vim}
 
-# Install TYPO3 using Composer
-cd /var/www/html
-echo "Downloading TYPO3 source"
-sudo -u apache composer create-project typo3/cms-base-distribution:^11 typo3 --no-interaction --no-progress
+  # SELinux serve files off Apache, resursive
+  mkdir /var/www/html/typo3
+  chown apache:apache /var/www/html/typo3
 
-cd /var/www/html/typo3
-sudo -u apache composer require typo3/cms-lowlevel:^11 typo3/cms-introduction --no-interaction --no-progress
+  # Configure SELinux
+  sudo chcon -t httpd_sys_rw_content_t /var/www/html/typo3 -R # change context
+  sudo setsebool -P httpd_can_network_connect 1 # allow network connections
 
-# Redis specific
-sudo -u apache composer require b13/distributed-locks --no-interaction --no-progress
+  # Install TYPO3 using Composer
+  cd /var/www/html
+  echo "Downloading TYPO3 source"
+  sudo -u apache composer create-project typo3/cms-base-distribution:^11 typo3 --no-interaction --no-progress
 
-# Dotenv specific
-sudo -u apache composer config --no-plugins allow-plugins.helhum/dotenv-connector true
-sudo -u apache composer require helhum/dotenv-connector --no-interaction --no-progress
-sudo -u apache composer dumpautoload
+  cd /var/www/html/typo3
+  sudo -u apache composer require typo3/cms-lowlevel:^11 typo3/cms-introduction --no-interaction --no-progress
 
-# # Create an empty FIRST_INSTALL file to enable TYPO3 install tool
-sudo -u apache touch /var/www/html/typo3/public/FIRST_INSTALL
+  # Redis specific
+  sudo -u apache composer require b13/distributed-locks --no-interaction --no-progress
+
+  # Dotenv specific
+  sudo -u apache composer config --no-plugins allow-plugins.helhum/dotenv-connector true
+  sudo -u apache composer require helhum/dotenv-connector --no-interaction --no-progress
+  sudo -u apache composer dumpautoload
+
+  # # Create an empty FIRST_INSTALL file to enable TYPO3 install tool
+  sudo -u apache touch /var/www/html/typo3/public/FIRST_INSTALL
+
+  touch "$FLAG_FILE"
+fi
